added kubeconfig download feature and corresponding endpoint

+++ backend/src/websocket.js
@@ -141,6 +141,9 @@ const connectionCheck = function (transactionId, nodes, wsClient, dryRun) {
 
         // Exec ansible connection test
         if (!Boolean(dryRun)) {
+
+            wsClient.dryRun = true;
+
             const options = {
                 cwd: dir,
                 env: null
@@ -166,7 +169,6 @@ const connectionCheck = function (transactionId, nodes, wsClient, dryRun) {
                 if (stdout) {
                     const ansibleFacts = JSON.parse(stdout).plays[0].tasks[0].hosts;
                     const ansibleStats = JSON.parse(stdout).stats;
-                    console.log('stdout:');
                     nodes.forEach((node) => {
                         // console.log(stdoutJson[node.host]);
                         if (ansibleStats[node.host].unreachable) {
@@ -225,8 +227,6 @@ const connectionCheck = function (transactionId, nodes, wsClient, dryRun) {
                     });
                 }
                 if (stderr) {
-                    console.log('stderr:');
-                    console.log(stderr);
                     wsMsg.event = 'ERROR';
                     wsMsg.payload = JSON.stringify(stderr);
                     sendMessage(wsMsg, wsClient);
@@ -277,13 +277,15 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
         payload: ''
     }
 
-    console.log('Dry-Run Mode', dryRun)
-
     // Send init msg through ws
     // sendMessage(wsMsg, wsClient);
     const dir = `${basePath}/${wsClient.uuid}`;
     try {
         if (!Boolean(dryRun)) {
+            console.log(`${transactionId} Setup started`);
+
+            // create ansible env
+            process.env.ANSIBLE_HOST_KEY_CHECKING = false;
             fs.mkdirSync(`${dir}/group_vars/all`, {recursive: true});
             fs.copyFileSync(`${basePath}/group_vars.testing/all/cert-manager.yaml`, `${dir}/group_vars/all/cert-manager.yaml`);
             fs.copyFileSync(`${basePath}/group_vars.testing/all/global.yaml`, `${dir}/group_vars/all/global.yaml`);
@@ -303,7 +305,7 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
 
             // Setting up ENV
             // process.env.ANSIBLE_TIMEOUT = 3;
-            process.env.ANSIBLE_HOST_KEY_CHECKING = false;
+
             // process.env.ANSIBLE_LOAD_CALLBACK_PLUGINS = true;
             // process.env.ANSIBLE_STDOUT_CALLBACK = 'json';
 
@@ -334,8 +336,8 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
             });
 
             ans.on('close', code => {
-                // todo: read kubeconfig kubeadm.conf
-                const kubeConf = fs.readFileSync(`${dir}/kubeadm.conf`, 'utf8');
+                // todo: fail safety missing
+                wsClient.setup = fs.readFileSync(`${dir}/kubeadm.conf`, 'utf8');
                 wsClient['kubeConfig'] = kubeConf;
                 wsMsg.event = 'EXECUTION';
                 wsMsg.payload = kubeConf;
@@ -348,7 +350,10 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
             });
 
         } else {
+
             console.log(`${transactionId} Setup running in dry-run-mode`);
+            wsClient.dryRun = true;
+
             const dryRunScriptsPath = path.join(__dirname, 'templates');
             const options = {
                 cwd: dryRunScriptsPath,
@@ -383,11 +388,9 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
 
             dryExec.on('close', code => {
                 // todo: read kubeconfig kubeadm.conf
-                // const kubeConf = fs.readFileSync(`${dir}/kubeadm.conf`, 'utf8');
-                // wsClient['kubeConfig'] = kubeConf;
-                // wsMsg.event = 'EXECUTION';
-                // wsMsg.payload = kubeConf;
-                // sendMessage(wsMsg, wsClient);
+                wsClient.setup = fs.readFileSync(path.join(__dirname,
+                    'templates/kube.config.template'),
+                    'utf8');
                 console.log(code);
                 wsMsg.event = 'DONE';
                 wsMsg.payload = code;
@@ -408,6 +411,17 @@ const setup = function (transactionId, basePath, dryRun, wsClient, hostsYaml) {
         // Cleanup
         cleanUpPath(transactionId, dir, ['hosts.yml', 'key.pem']);
     }
+}
+
+const downloadKubeConf = function (client) {
+    const dir = `/tmp/${client.uuid}`;
+    try {
+        fs.mkdirSync(dir, {recursive: true});
+        fs.writeFileSync(`${dir}/kubeconfig`, client.setup);
+    } catch (e) {
+        console.log(e);
+    }
+    return `${dir}`
 }
 
 // invoke like so: randPassword(5,3,2);
@@ -451,19 +465,21 @@ const cleanUpPath = function (transactionId, baseDir, files) {
         process.env.ANSIBLE_LOAD_CALLBACK_PLUGINS = '';
         process.env.ANSIBLE_STDOUT_CALLBACK = '';
         if (fs.existsSync(baseDir)) {
-            files.forEach((file) => {
-                const filePath = path.join(baseDir, file);
-                if (fs.existsSync(filePath)) {
-                    fs.unlinkSync(filePath);
-                } else {
-                    console.error(`${file} does not exist`)
-                }
-            });
+            if (files != null) {
+                files.forEach((file) => {
+                    const filePath = path.join(baseDir, file);
+                    if (fs.existsSync(filePath)) {
+                        fs.unlinkSync(filePath);
+                    } else {
+                        console.error(`${file} does not exist`)
+                    }
+                });
+            }
             fs.rmdirSync(baseDir, {recursive: true});
         } else {
             console.error(`${baseDir} does not exist`)
         }
-        console.log(`${transactionId} Cleanup done`);
+        console.log(`${(transactionId == null) ? '' : transactionId} Cleanup done`);
     } catch (cleaningError) {
         console.log('Cleaning up gone wrong: ', cleaningError);
     }
@@ -487,5 +503,7 @@ module.exports = {
     setup,
     sleep,
     genTransactionId,
-    randPassword
+    randPassword,
+    downloadKubeConf,
+    cleanUpPath
 };
